moving all the project to the src directory
