+++ .gitignore
@@ -1,6 +1,4 @@
-last.txt
-languages.txt
-older
 venv
 configurations.txt
 .idea
+src/resources
