--- init.py
@@ -1,4 +1,0 @@
-#!/usr/bin/python
-from argumentParsing import configurations
-
-configurations.init()
