+++ src/translating/web/wrongStatusCodeException.py
